improvement(components): change service section according to design
