+++ src/component/service-section/service-card.astro
@@ -1,0 +1,36 @@
+---
+import type { ServiceCardProps } from "./service-card.type";
+
+interface Props extends ServiceCardProps {}
+
+const { body, hoverLightingColor, href, iconPath, title, iconBgColor } =
+  Astro.props;
+---
+
+<a
+  href={href}
+  class="shadow-section group transition ease-in-out hover:bg-gray-900 bg-neutral-50 rounded"
+>
+  <div class="w-full h-full space-y-8 relative overflow-clip p-8">
+    <div
+      class={`p-3 ${iconBgColor} inline-block group-hover:bg-neutral-50 rounded`}
+    >
+      <img src={iconPath} alt="Card Icon" />
+    </div>
+    <h3 class="text-slate-700 text-2xl font-bold group-hover:text-neutral-50">
+      {title}
+    </h3>
+    <p class="text-slate-700 text-base group-hover:text-neutral-50">{body}</p>
+    <button class="text-blue-700 group-hover:text-neutral-50">Learn more</button
+    >
+
+    <div
+      class={`glow-def absolute -left-6 -top-10 hidden h-20 w-20 rounded-full blur-3xl group-hover:block ${hoverLightingColor}`}
+    >
+    </div>
+    <div
+      class={`glow-def absolute -bottom-10 -right-9 hidden h-20 w-20 rounded-full blur-3xl group-hover:block ${hoverLightingColor}`}
+    >
+    </div>
+  </div>
+</a>
