+++ src/component/service-section/service-card.type.ts
@@ -1,0 +1,8 @@
+export type ServiceCardProps = {
+  iconPath: string;
+  title: string;
+  body: string;
+  href: string;
+  hoverLightingColor: string;
+  iconBgColor: string;
+};
